runpod_server/requirements.txt
# MCP Server
mcp>=0.1.0

# DeepSeek-VL2
transformers>=4.40.0
torch>=2.0.0
pillow>=10.0.0
accelerate>=0.20.0
sentencepiece>=0.1.99

# Utilities
pyyaml>=6.0
